# perf(clean): improve performance of clean_duplication