# Description

Improves the performance of the clean_duplication function. The improvement comes from using parallelizing with dask and also only adding the count of cluster entries and sorting by the counts on page load. 

Some rough results on my machine:
N-gram clustering a million rows went from ~90 seconds to ~60 seconds. When there is lots of clusters the improvement is not massive.
Fingerprint clustering a million rows went from ~40 seconds to ~7 seconds. This resulted in no clusters which is when improvement is largest.
Fingerprint clustering 200,000 rows went from ~0.7 seconds to ~0.5 seconds. This dataset has a lot of duplicates which results in much faster times but the difference is minimal.

For the first two results I just used `df = pd.DataFrame({"data": [str(i) for i in range(1_000_000)]})`. The last result uses this dataset https://www.kaggle.com/tamber/steam-video-games.


# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have already squashed the commits and make the commit message conform to the project standard.
- [x] I have already marked the commit with "BREAKING CHANGE" or "Fixes #" if needed.
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules