@ryanwdale 2021-04-12
Thank you for the excellent comments @brandonlockhart. I made all the changes and the performance improved a lot, it's down to ~40 seconds for n-gram clustering a million rows.